fix(shared): moved tags and current-projects template data
